requirements.txt
requests
simplejson
beautifulsoup4
openai
notion-client
python-dotenv
